Mainbank

+++ lib/Pages/qr_code_point.dart
@@ -3,12 +3,12 @@ import 'dart:developer' as dev;
 import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
-import 'package:hajeri_demo/Pages/display_qr.dart';
-import 'package:hajeri_demo/Pages/generate_qr.dart';
-import 'package:hajeri_demo/Pages/maintain_qr.dart';
-import 'package:hajeri_demo/constant.dart';
-import 'package:hajeri_demo/main.dart';
-import 'package:hajeri_demo/url.dart';
+import '../Pages/display_qr.dart';
+import '../Pages/generate_qr.dart';
+import '../Pages/maintain_qr.dart';
+import '../constant.dart';
+import '../main.dart';
+import '../url.dart';
 import 'package:qr_flutter/qr_flutter.dart';
 import 'package:http/http.dart' as http;
 
@@ -186,13 +186,7 @@ class _QrCodePointViewState extends State<QrCodePointView> {
                                 color: Colors.white,
                               ),
                               onPressed: () async {
-                                String result = await deleteQrCodePoint(
-                                  mobile: qrCodePoint['mobile'],
-                                  point: qrCodePoint['nameofqrcodepoint'],
-                                );
-
-                                if (result != null) {
-                                  showDialog(
+                                bool delete = await showDialog(
                                     context: context,
                                     barrierDismissible: false,
                                     builder: (context) {
@@ -200,20 +194,56 @@ class _QrCodePointViewState extends State<QrCodePointView> {
                                         actions: [
                                           TextButton(
                                             onPressed: () {
-                                              Navigator.pushNamed(
-                                                context,
-                                                MaintainQr.id,
-                                              );
+                                              Navigator.of(context).pop(true);
                                             },
-                                            child: Text('back'),
+                                            child: Text('Yes'),
+                                          ),
+                                          TextButton(
+                                            onPressed: () {
+                                              Navigator.of(context).pop(false);
+                                            },
+                                            child: Text('No'),
                                           )
                                         ],
+                                        title: Text('Delete Qr'),
                                         content: Text(
-                                          result,
+                                          'Do you want to delete qr code point ?',
                                         ),
                                       );
-                                    },
+                                    });
+                                if (delete) {
+                                  String result = await deleteQrCodePoint(
+                                    mobile: qrCodePoint['mobile'],
+                                    point: qrCodePoint['nameofqrcodepoint'],
                                   );
+
+                                  if (result != null) {
+                                    showDialog(
+                                      context: context,
+                                      barrierDismissible: false,
+                                      builder: (context) {
+                                        return AlertDialog(
+                                          actions: [
+                                            TextButton(
+                                              onPressed: () {
+                                                Navigator.push(
+                                                    context,
+                                                    MaterialPageRoute(
+                                                      builder: (context) =>
+                                                          ShowCaseQr(),
+                                                    ));
+                                              },
+                                              child: Text('back'),
+                                            )
+                                          ],
+                                          title: Text('Delete Qr'),
+                                          content: Text(
+                                            result,
+                                          ),
+                                        );
+                                      },
+                                    );
+                                  }
                                 }
                               },
                             ),
@@ -227,7 +257,7 @@ class _QrCodePointViewState extends State<QrCodePointView> {
             ),
           ),
           DraggableScrollableSheet(
-            initialChildSize: 0.7,
+            initialChildSize: 0.4,
             minChildSize: 0.35,
             maxChildSize: 0.8,
             builder: (BuildContext context, ScrollController scrollController) {
@@ -327,13 +357,41 @@ class _QrCodePointViewState extends State<QrCodePointView> {
                                     color: Colors.white,
                                   ),
                                   onTap: () async {
-                                    String result = await deleteQrCodePoint(
-                                      mobile: qrCodePoint['mobile'],
-                                      point: qrCodePoint['nameofqrcodepoint'],
-                                    );
+                                    bool delete = await showDialog(
+                                        context: context,
+                                        barrierDismissible: false,
+                                        builder: (context) {
+                                          return AlertDialog(
+                                            actions: [
+                                              TextButton(
+                                                onPressed: () {
+                                                  Navigator.of(context)
+                                                      .pop(true);
+                                                },
+                                                child: Text('Yes'),
+                                              ),
+                                              TextButton(
+                                                onPressed: () {
+                                                  Navigator.of(context)
+                                                      .pop(false);
+                                                },
+                                                child: Text('No'),
+                                              )
+                                            ],
+                                            title: Text('Delete Qr'),
+                                            content: Text(
+                                              'Do you want to delete qr code point ?',
+                                            ),
+                                          );
+                                        });
+                                    if (delete) {
+                                      String result = await deleteQrCodePoint(
+                                        mobile: qrCodePoint['mobile'],
+                                        point: qrCodePoint['nameofqrcodepoint'],
+                                      );
 
-                                    if (result != null) {
-                                      showDialog(
+                                      if (result != null) {
+                                        showDialog(
                                           context: context,
                                           barrierDismissible: false,
                                           builder: (context) {
@@ -341,19 +399,24 @@ class _QrCodePointViewState extends State<QrCodePointView> {
                                               actions: [
                                                 TextButton(
                                                   onPressed: () {
-                                                    Navigator.pushNamed(
-                                                      context,
-                                                      MaintainQr.id,
-                                                    );
+                                                    Navigator.push(
+                                                        context,
+                                                        MaterialPageRoute(
+                                                          builder: (context) =>
+                                                              ShowCaseQr(),
+                                                        ));
                                                   },
                                                   child: Text('back'),
                                                 )
                                               ],
+                                              title: Text('Delete Qr'),
                                               content: Text(
                                                 result,
                                               ),
                                             );
-                                          });
+                                          },
+                                        );
+                                      }
                                     }
                                   },
                                   decoration: BoxDecoration(
